refactor: moved out flame utilities
