+++ stories/icicle/02_unix_flame.tsx
@@ -20,88 +20,23 @@
 import React from 'react';
 
 import { Chart, Datum, Partition, PartitionLayout, Settings } from '../../src';
-import { ShapeTreeNode } from '../../src/chart_types/partition_chart/layout/types/viewmodel_types';
-import { PrimitiveValue } from '../../src/chart_types/partition_chart/layout/utils/group_by_rollup';
-import { mocks } from '../../src/mocks/hierarchical';
 import { STORYBOOK_LIGHT_THEME } from '../shared';
-import { categoricalFillColor, plasma18 as palette } from '../utils/utils';
+import { config, getFlatData, getLayerSpec, maxDepth } from '../utils/hierarchical_input_utils';
+import { plasma18 as palette } from '../utils/utils';
 
 const color = palette.slice().reverse();
-
-const raw = mocks.observabilityTree;
-
-interface Node {
-  c?: Node[];
-  n: string;
-  v: number;
-}
-
-type Row = { [layerKey: string]: unknown; value: number; depth: number };
-
-const flatTree = ({ c, n, v }: Node, depth: number): Row[] => {
-  if (!c) {
-    return [{ [`layer_${depth}`]: n, value: v, depth }];
-  }
-  // looks like our test runner can't run c.flatMap(...)
-  const childrenRows = c.reduce<Row[]>((a, child) => [...a, ...flatTree(child, depth + 1)], []);
-  const childrenTotal = childrenRows.reduce((p, { value }) => p + value, 0);
-  const missing = Math.max(0, v - childrenTotal);
-  if (missing > 0) {
-    childrenRows.unshift({ [`layer_${depth + 1}`]: undefined, value: missing / 2, depth });
-    childrenRows.push({ [`layer_${depth + 1}`]: undefined, value: missing / 2, depth });
-  }
-  childrenRows.forEach((innerChild) => {
-    innerChild[`layer_${depth}`] = n;
-  });
-  return childrenRows;
-};
-
-const flatData = flatTree(raw, 0);
-const maxDepth = flatData.reduce((p, n) => Math.max(p, n.depth), 0);
-
-const layerSpec = [...new Array(maxDepth + 1)].map((_, depth) => ({
-  groupByRollup: (d: Datum) => d[`layer_${depth}`],
-  nodeLabel: (d: PrimitiveValue) => String(d),
-  showAccessor: (d: PrimitiveValue) => d !== undefined,
-  shape: {
-    fillColor: (d: ShapeTreeNode) => (d.dataName ? categoricalFillColor(color, 0.8)(depth) : 'transparent'),
-  },
-}));
 
 export const Example = () => {
   return (
     <Chart className="story-chart" size={{ height: 512 }}>
-      <Settings showLegend flatLegend legendMaxDepth={18} theme={STORYBOOK_LIGHT_THEME} />
+      <Settings showLegend flatLegend legendMaxDepth={maxDepth} theme={STORYBOOK_LIGHT_THEME} />
       <Partition
         id="spec_1"
-        data={flatData}
+        data={getFlatData()}
         valueAccessor={(d: Datum) => d.value as number}
         valueFormatter={() => ''}
-        layers={layerSpec}
-        config={{
-          partitionLayout: PartitionLayout.flame,
-          fontFamily: 'Arial',
-          fillLabel: {
-            valueFormatter: (d: number) => d,
-            // fontStyle: 'italic',
-            textInvertible: true,
-            // fontWeight: 900,
-            valueFont: {
-              // fontFamily: 'Menlo',
-              // fontStyle: 'normal',
-              // fontWeight: 100,
-            },
-          },
-          margin: { top: 0, bottom: 0, left: 0, right: 0 },
-          minFontSize: 8,
-          maxFontSize: 14,
-          idealFontSizeJump: 1.01,
-          outerSizeRatio: 1,
-          emptySizeRatio: 0,
-          circlePadding: 0,
-          radialPadding: 0,
-          backgroundColor: 'rgba(229,229,229,1)',
-        }}
+        layers={getLayerSpec(color)}
+        config={{ ...config, partitionLayout: PartitionLayout.flame }}
       />
     </Chart>
   );
