+++ stories/stylings/20_partition_background.tsx
@@ -25,7 +25,7 @@ import { config } from '../../src/chart_types/partition_chart/layout/config/conf
 import { ShapeTreeNode } from '../../src/chart_types/partition_chart/layout/types/viewmodel_types';
 import { mocks } from '../../src/mocks/hierarchical';
 import {
-  categoricalFillColor,
+  discreteColor,
   colorBrewerCategoricalStark9,
   countryLookup,
   productLookup,
@@ -55,15 +55,14 @@ export const Example = () => {
             groupByRollup: (d: Datum) => d.sitc1,
             nodeLabel: (d: any) => productLookup[d].name,
             shape: {
-              fillColor: (d: ShapeTreeNode) => categoricalFillColor(colorBrewerCategoricalStark9, 0.7)(d.sortIndex),
+              fillColor: (d: ShapeTreeNode) => discreteColor(colorBrewerCategoricalStark9, 0.7)(d.sortIndex),
             },
           },
           {
             groupByRollup: (d: Datum) => countryLookup[d.dest].continentCountry.slice(0, 2),
             nodeLabel: (d: any) => regionLookup[d].regionName,
             shape: {
-              fillColor: (d: ShapeTreeNode) =>
-                categoricalFillColor(colorBrewerCategoricalStark9, 0.5)(d.parent.sortIndex),
+              fillColor: (d: ShapeTreeNode) => discreteColor(colorBrewerCategoricalStark9, 0.5)(d.parent.sortIndex),
             },
           },
           {
@@ -71,7 +70,7 @@ export const Example = () => {
             nodeLabel: (d: any) => countryLookup[d].name,
             shape: {
               fillColor: (d: ShapeTreeNode) =>
-                categoricalFillColor(colorBrewerCategoricalStark9, 0.3)(d.parent.parent.sortIndex),
+                discreteColor(colorBrewerCategoricalStark9, 0.3)(d.parent.parent.sortIndex),
             },
           },
         ]}
