+++ stories/sunburst/9_sunburst_three_layers.tsx
@@ -26,7 +26,7 @@ import { ShapeTreeNode } from '../../src/chart_types/partition_chart/layout/type
 import { mocks } from '../../src/mocks/hierarchical';
 import { STORYBOOK_LIGHT_THEME } from '../shared';
 import {
-  categoricalFillColor,
+  discreteColor,
   colorBrewerCategoricalStark9,
   countryLookup,
   productLookup,
@@ -47,7 +47,7 @@ export const Example = () => (
           nodeLabel: (d: any) => productLookup[d].name,
           fillLabel: { maximizeFontSize: boolean('Maximize font size layer 1', true) },
           shape: {
-            fillColor: (d: ShapeTreeNode) => categoricalFillColor(colorBrewerCategoricalStark9, 0.7)(d.sortIndex),
+            fillColor: (d: ShapeTreeNode) => discreteColor(colorBrewerCategoricalStark9, 0.7)(d.sortIndex),
           },
         },
         {
@@ -55,8 +55,7 @@ export const Example = () => (
           nodeLabel: (d: any) => regionLookup[d].regionName,
           fillLabel: { maximizeFontSize: boolean('Maximize font size layer 2', true) },
           shape: {
-            fillColor: (d: ShapeTreeNode) =>
-              categoricalFillColor(colorBrewerCategoricalStark9, 0.5)(d.parent.sortIndex),
+            fillColor: (d: ShapeTreeNode) => discreteColor(colorBrewerCategoricalStark9, 0.5)(d.parent.sortIndex),
           },
         },
         {
@@ -65,7 +64,7 @@ export const Example = () => (
           fillLabel: { maximizeFontSize: boolean('Maximize font size layer 3', true) },
           shape: {
             fillColor: (d: ShapeTreeNode) =>
-              categoricalFillColor(colorBrewerCategoricalStark9, 0.3)(d.parent.parent.sortIndex),
+              discreteColor(colorBrewerCategoricalStark9, 0.3)(d.parent.parent.sortIndex),
           },
         },
       ]}
