+++ stories/utils/utils.ts
@@ -195,7 +195,7 @@ export const colorBrewerCategoricalStark9: RGBStrings = [
   'rgb(153,153,153)',
 ].map(rgbStringToTuple) as RGBStrings;
 
-export const categoricalFillColor = (categoricalColors: RGBStrings, opacity = 1) => (i: number) =>
+export const discreteColor = (categoricalColors: RGBStrings, opacity = 1) => (i: number) =>
   `rgba(${categoricalColors[i % categoricalColors.length].concat([opacity.toString()]).join(',')})`;
 
 export const decreasingOpacityCET2 = (opacity: number) => (d: any, i: number, a: any[]) =>
